Move tests regarding discount in cart to separate file
